Score samples using the ASCI tool

Score samples using the ASCI tool

ASCI(taxa, stations)

Arguments

taxa
data.frame for input taxonomy data

stations
data.frame for input station data

Value

A dataframe with all metrics calculated for each provided taxa

Details

One index for three taxonomy types are scored, MMI for diatoms, soft-bodied algae, and hybrid. If only soft-bodied algae or diatoms are present, only the respective ASCI and metrics are returned.

Examples

# calculate all
ASCI(demo_algae_tax, demo_station)

#> # A tibble: 3 x 109
#>   SampleID StationCode SampleDate          Replicate SampleType D_ValveCount
#>   <chr>    <chr>       <dttm>                  <dbl> <chr>             <int>
#> 1 404M073… 404M07357   2016-06-13 00:00:00         1 Integrate…          600
#> 2 801M169… 801M16916   2016-05-25 00:00:00         1 Microalga…          600
#> 3 909M249… 909M24937   2016-06-22 00:00:00         1 Macroalga…          600
#> # … with 103 more variables: S_EntityCount <int>, S_Biovolume <dbl>,
#> #   D_NumberTaxa <int>, S_NumberTaxa <int>, H_NumberTaxa <int>,
#> #   UnrecognizedTaxa <chr>, D_ASCI <dbl>, S_ASCI <dbl>, H_ASCI <dbl>,
#> #   D_cnt.spp.most.tol_mod <dbl>, D_cnt.spp.most.tol_mod_scr <dbl>,
#> #   D_cnt.spp.most.tol_pct_att <dbl>, D_cnt.spp.most.tol_pred <dbl>,
#> #   D_cnt.spp.most.tol_raw <int>, D_cnt.spp.most.tol_raw_scr <dbl>,
#> #   D_EpiRho.richness_mod <dbl>, D_EpiRho.richness_mod_scr <dbl>,
#> #   D_EpiRho.richness_pct_att <dbl>, D_EpiRho.richness_pred <dbl>,
#> #   D_EpiRho.richness_raw <int>, D_EpiRho.richness_raw_scr <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_mod <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_mod_scr <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_pct_att <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_pred <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_raw <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_raw_scr <dbl>,
#> #   D_prop.spp.Planktonic_mod <dbl>, D_prop.spp.Planktonic_mod_scr <dbl>,
#> #   D_prop.spp.Planktonic_pct_att <dbl>, D_prop.spp.Planktonic_pred <dbl>,
#> #   D_prop.spp.Planktonic_raw <dbl>, D_prop.spp.Planktonic_raw_scr <dbl>,
#> #   D_prop.spp.Trophic.E_mod <dbl>, D_prop.spp.Trophic.E_mod_scr <dbl>,
#> #   D_prop.spp.Trophic.E_pct_att <dbl>, D_prop.spp.Trophic.E_pred <dbl>,
#> #   D_prop.spp.Trophic.E_raw <dbl>, D_prop.spp.Trophic.E_raw_scr <dbl>,
#> #   D_Salinity.BF.richness_mod <dbl>, D_Salinity.BF.richness_mod_scr <dbl>,
#> #   D_Salinity.BF.richness_pct_att <dbl>, D_Salinity.BF.richness_pred <dbl>,
#> #   D_Salinity.BF.richness_raw <int>, D_Salinity.BF.richness_raw_scr <dbl>,
#> #   S_prop.spp.IndicatorClass_DOC_high_raw_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_DOC_high_raw_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_raw_scr <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_raw_scr <dbl>,
#> #   S_prop.spp.ZHR_pct_att <dbl>, S_prop.spp.ZHR_raw <dbl>,
#> #   S_prop.spp.ZHR_raw_scr <dbl>, H_cnt.spp.IndicatorClass_TP_high_mod <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_mod_scr <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_pct_att <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_pred <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_raw <int>,
#> #   H_cnt.spp.IndicatorClass_TP_high_raw_scr <dbl>,
#> #   H_cnt.spp.most.tol_mod <dbl>, H_cnt.spp.most.tol_mod_scr <dbl>,
#> #   H_cnt.spp.most.tol_pct_att <dbl>, H_cnt.spp.most.tol_pred <dbl>,
#> #   H_cnt.spp.most.tol_raw <int>, H_cnt.spp.most.tol_raw_scr <dbl>,
#> #   H_EpiRho.richness_mod <dbl>, H_EpiRho.richness_mod_scr <dbl>,
#> #   H_EpiRho.richness_pct_att <dbl>, H_EpiRho.richness_pred <dbl>,
#> #   H_EpiRho.richness_raw <int>, H_EpiRho.richness_raw_scr <dbl>,
#> #   H_OxyRed.DO_30.richness_mod <dbl>, H_OxyRed.DO_30.richness_mod_scr <dbl>,
#> #   H_OxyRed.DO_30.richness_pct_att <dbl>, H_OxyRed.DO_30.richness_pred <dbl>,
#> #   H_OxyRed.DO_30.richness_raw <int>, H_OxyRed.DO_30.richness_raw_scr <dbl>,
#> #   H_prop.spp.Planktonic_mod <dbl>, H_prop.spp.Planktonic_mod_scr <dbl>,
#> #   H_prop.spp.Planktonic_pct_att <dbl>, H_prop.spp.Planktonic_pred <dbl>,
#> #   H_prop.spp.Planktonic_raw <dbl>, H_prop.spp.Planktonic_raw_scr <dbl>,
#> #   H_prop.spp.Trophic.E_mod <dbl>, H_prop.spp.Trophic.E_mod_scr <dbl>,
#> #   H_prop.spp.Trophic.E_pct_att <dbl>, H_prop.spp.Trophic.E_pred <dbl>,
#> #   H_prop.spp.Trophic.E_raw <dbl>, H_prop.spp.Trophic.E_raw_scr <dbl>,
#> #   H_prop.spp.ZHR_pct_att <dbl>, H_prop.spp.ZHR_raw <dbl>,
#> #   H_prop.spp.ZHR_raw_scr <dbl>, H_Salinity.BF.richness_mod <dbl>,
#> #   H_Salinity.BF.richness_mod_scr <dbl>, H_Salinity.BF.richness_pct_att <dbl>,
#> #   H_Salinity.BF.richness_pred <dbl>, H_Salinity.BF.richness_raw <int>, …

# works if either soft-bodied or diatoms are missing
# remove diatoms from station sample 909M24937
tmp <- subset(demo_algae_tax, !(StationCode == '909M24937' & SampleTypeCode == 'Integrated'))
ASCI(tmp, demo_station)

#> Warning: Only diatom or soft-bodied algae present at sites: 909M24937_2016-06-22_1

#> # A tibble: 2 x 109
#>   SampleID StationCode SampleDate          Replicate SampleType D_ValveCount
#>   <chr>    <chr>       <dttm>                  <dbl> <chr>             <int>
#> 1 404M073… 404M07357   2016-06-13 00:00:00         1 Integrate…          600
#> 2 801M169… 801M16916   2016-05-25 00:00:00         1 Microalga…          600
#> # … with 103 more variables: S_EntityCount <int>, S_Biovolume <dbl>,
#> #   D_NumberTaxa <int>, S_NumberTaxa <int>, H_NumberTaxa <int>,
#> #   UnrecognizedTaxa <chr>, D_ASCI <dbl>, S_ASCI <dbl>, H_ASCI <dbl>,
#> #   D_cnt.spp.most.tol_mod <dbl>, D_cnt.spp.most.tol_mod_scr <dbl>,
#> #   D_cnt.spp.most.tol_pct_att <dbl>, D_cnt.spp.most.tol_pred <dbl>,
#> #   D_cnt.spp.most.tol_raw <int>, D_cnt.spp.most.tol_raw_scr <dbl>,
#> #   D_EpiRho.richness_mod <dbl>, D_EpiRho.richness_mod_scr <dbl>,
#> #   D_EpiRho.richness_pct_att <dbl>, D_EpiRho.richness_pred <dbl>,
#> #   D_EpiRho.richness_raw <int>, D_EpiRho.richness_raw_scr <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_mod <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_mod_scr <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_pct_att <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_pred <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_raw <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_raw_scr <dbl>,
#> #   D_prop.spp.Planktonic_mod <dbl>, D_prop.spp.Planktonic_mod_scr <dbl>,
#> #   D_prop.spp.Planktonic_pct_att <dbl>, D_prop.spp.Planktonic_pred <dbl>,
#> #   D_prop.spp.Planktonic_raw <dbl>, D_prop.spp.Planktonic_raw_scr <dbl>,
#> #   D_prop.spp.Trophic.E_mod <dbl>, D_prop.spp.Trophic.E_mod_scr <dbl>,
#> #   D_prop.spp.Trophic.E_pct_att <dbl>, D_prop.spp.Trophic.E_pred <dbl>,
#> #   D_prop.spp.Trophic.E_raw <dbl>, D_prop.spp.Trophic.E_raw_scr <dbl>,
#> #   D_Salinity.BF.richness_mod <dbl>, D_Salinity.BF.richness_mod_scr <dbl>,
#> #   D_Salinity.BF.richness_pct_att <dbl>, D_Salinity.BF.richness_pred <dbl>,
#> #   D_Salinity.BF.richness_raw <int>, D_Salinity.BF.richness_raw_scr <dbl>,
#> #   S_prop.spp.IndicatorClass_DOC_high_raw_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_DOC_high_raw_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_raw_scr <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_raw_scr <dbl>,
#> #   S_prop.spp.ZHR_pct_att <dbl>, S_prop.spp.ZHR_raw <dbl>,
#> #   S_prop.spp.ZHR_raw_scr <dbl>, H_cnt.spp.IndicatorClass_TP_high_mod <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_mod_scr <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_pct_att <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_pred <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_raw <int>,
#> #   H_cnt.spp.IndicatorClass_TP_high_raw_scr <dbl>,
#> #   H_cnt.spp.most.tol_mod <dbl>, H_cnt.spp.most.tol_mod_scr <dbl>,
#> #   H_cnt.spp.most.tol_pct_att <dbl>, H_cnt.spp.most.tol_pred <dbl>,
#> #   H_cnt.spp.most.tol_raw <int>, H_cnt.spp.most.tol_raw_scr <dbl>,
#> #   H_EpiRho.richness_mod <dbl>, H_EpiRho.richness_mod_scr <dbl>,
#> #   H_EpiRho.richness_pct_att <dbl>, H_EpiRho.richness_pred <dbl>,
#> #   H_EpiRho.richness_raw <int>, H_EpiRho.richness_raw_scr <dbl>,
#> #   H_OxyRed.DO_30.richness_mod <dbl>, H_OxyRed.DO_30.richness_mod_scr <dbl>,
#> #   H_OxyRed.DO_30.richness_pct_att <dbl>, H_OxyRed.DO_30.richness_pred <dbl>,
#> #   H_OxyRed.DO_30.richness_raw <int>, H_OxyRed.DO_30.richness_raw_scr <dbl>,
#> #   H_prop.spp.Planktonic_mod <dbl>, H_prop.spp.Planktonic_mod_scr <dbl>,
#> #   H_prop.spp.Planktonic_pct_att <dbl>, H_prop.spp.Planktonic_pred <dbl>,
#> #   H_prop.spp.Planktonic_raw <dbl>, H_prop.spp.Planktonic_raw_scr <dbl>,
#> #   H_prop.spp.Trophic.E_mod <dbl>, H_prop.spp.Trophic.E_mod_scr <dbl>,
#> #   H_prop.spp.Trophic.E_pct_att <dbl>, H_prop.spp.Trophic.E_pred <dbl>,
#> #   H_prop.spp.Trophic.E_raw <dbl>, H_prop.spp.Trophic.E_raw_scr <dbl>,
#> #   H_prop.spp.ZHR_pct_att <dbl>, H_prop.spp.ZHR_raw <dbl>,
#> #   H_prop.spp.ZHR_raw_scr <dbl>, H_Salinity.BF.richness_mod <dbl>,
#> #   H_Salinity.BF.richness_mod_scr <dbl>, H_Salinity.BF.richness_pct_att <dbl>,
#> #   H_Salinity.BF.richness_pred <dbl>, H_Salinity.BF.richness_raw <int>, …

# works if either soft-bodied or diatoms are missing
# remove soft-bodied from station sample 801M16916
tmp <- subset(demo_algae_tax, !(StationCode == '801M16916' & SampleTypeCode != 'Integrated'))
ASCI(tmp, demo_station)

#> Warning: Only diatom or soft-bodied algae present at sites: 801M16916_2016-05-25_1

#> # A tibble: 2 x 109
#>   SampleID StationCode SampleDate          Replicate SampleType D_ValveCount
#>   <chr>    <chr>       <dttm>                  <dbl> <chr>             <int>
#> 1 404M073… 404M07357   2016-06-13 00:00:00         1 Integrate…          600
#> 2 909M249… 909M24937   2016-06-22 00:00:00         1 Macroalga…          600
#> # … with 103 more variables: S_EntityCount <int>, S_Biovolume <dbl>,
#> #   D_NumberTaxa <int>, S_NumberTaxa <int>, H_NumberTaxa <int>,
#> #   UnrecognizedTaxa <chr>, D_ASCI <dbl>, S_ASCI <dbl>, H_ASCI <dbl>,
#> #   D_cnt.spp.most.tol_mod <dbl>, D_cnt.spp.most.tol_mod_scr <dbl>,
#> #   D_cnt.spp.most.tol_pct_att <dbl>, D_cnt.spp.most.tol_pred <dbl>,
#> #   D_cnt.spp.most.tol_raw <int>, D_cnt.spp.most.tol_raw_scr <dbl>,
#> #   D_EpiRho.richness_mod <dbl>, D_EpiRho.richness_mod_scr <dbl>,
#> #   D_EpiRho.richness_pct_att <dbl>, D_EpiRho.richness_pred <dbl>,
#> #   D_EpiRho.richness_raw <int>, D_EpiRho.richness_raw_scr <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_mod <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_mod_scr <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_pct_att <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_pred <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_raw <dbl>,
#> #   D_prop.spp.IndicatorClass_TN_low_raw_scr <dbl>,
#> #   D_prop.spp.Planktonic_mod <dbl>, D_prop.spp.Planktonic_mod_scr <dbl>,
#> #   D_prop.spp.Planktonic_pct_att <dbl>, D_prop.spp.Planktonic_pred <dbl>,
#> #   D_prop.spp.Planktonic_raw <dbl>, D_prop.spp.Planktonic_raw_scr <dbl>,
#> #   D_prop.spp.Trophic.E_mod <dbl>, D_prop.spp.Trophic.E_mod_scr <dbl>,
#> #   D_prop.spp.Trophic.E_pct_att <dbl>, D_prop.spp.Trophic.E_pred <dbl>,
#> #   D_prop.spp.Trophic.E_raw <dbl>, D_prop.spp.Trophic.E_raw_scr <dbl>,
#> #   D_Salinity.BF.richness_mod <dbl>, D_Salinity.BF.richness_mod_scr <dbl>,
#> #   D_Salinity.BF.richness_pct_att <dbl>, D_Salinity.BF.richness_pred <dbl>,
#> #   D_Salinity.BF.richness_raw <int>, D_Salinity.BF.richness_raw_scr <dbl>,
#> #   S_prop.spp.IndicatorClass_DOC_high_raw_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_DOC_high_raw_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_NonRef_raw_scr <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_pct_att <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_raw <dbl>,
#> #   S_prop.spp.IndicatorClass_TP_high_raw_scr <dbl>,
#> #   S_prop.spp.ZHR_pct_att <dbl>, S_prop.spp.ZHR_raw <dbl>,
#> #   S_prop.spp.ZHR_raw_scr <dbl>, H_cnt.spp.IndicatorClass_TP_high_mod <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_mod_scr <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_pct_att <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_pred <dbl>,
#> #   H_cnt.spp.IndicatorClass_TP_high_raw <int>,
#> #   H_cnt.spp.IndicatorClass_TP_high_raw_scr <dbl>,
#> #   H_cnt.spp.most.tol_mod <dbl>, H_cnt.spp.most.tol_mod_scr <dbl>,
#> #   H_cnt.spp.most.tol_pct_att <dbl>, H_cnt.spp.most.tol_pred <dbl>,
#> #   H_cnt.spp.most.tol_raw <int>, H_cnt.spp.most.tol_raw_scr <dbl>,
#> #   H_EpiRho.richness_mod <dbl>, H_EpiRho.richness_mod_scr <dbl>,
#> #   H_EpiRho.richness_pct_att <dbl>, H_EpiRho.richness_pred <dbl>,
#> #   H_EpiRho.richness_raw <int>, H_EpiRho.richness_raw_scr <dbl>,
#> #   H_OxyRed.DO_30.richness_mod <dbl>, H_OxyRed.DO_30.richness_mod_scr <dbl>,
#> #   H_OxyRed.DO_30.richness_pct_att <dbl>, H_OxyRed.DO_30.richness_pred <dbl>,
#> #   H_OxyRed.DO_30.richness_raw <int>, H_OxyRed.DO_30.richness_raw_scr <dbl>,
#> #   H_prop.spp.Planktonic_mod <dbl>, H_prop.spp.Planktonic_mod_scr <dbl>,
#> #   H_prop.spp.Planktonic_pct_att <dbl>, H_prop.spp.Planktonic_pred <dbl>,
#> #   H_prop.spp.Planktonic_raw <dbl>, H_prop.spp.Planktonic_raw_scr <dbl>,
#> #   H_prop.spp.Trophic.E_mod <dbl>, H_prop.spp.Trophic.E_mod_scr <dbl>,
#> #   H_prop.spp.Trophic.E_pct_att <dbl>, H_prop.spp.Trophic.E_pred <dbl>,
#> #   H_prop.spp.Trophic.E_raw <dbl>, H_prop.spp.Trophic.E_raw_scr <dbl>,
#> #   H_prop.spp.ZHR_pct_att <dbl>, H_prop.spp.ZHR_raw <dbl>,
#> #   H_prop.spp.ZHR_raw_scr <dbl>, H_Salinity.BF.richness_mod <dbl>,
#> #   H_Salinity.BF.richness_mod_scr <dbl>, H_Salinity.BF.richness_pct_att <dbl>,
#> #   H_Salinity.BF.richness_pred <dbl>, H_Salinity.BF.richness_raw <int>, …